backend/requirements.txt
annotated-types==0.7.0
arrow==1.3.0
attrs==25.1.0
aws-lambda-wsgi==0.0.6
aws-sam-cli==1.133.0
aws-sam-translator==1.94.0
aws_lambda_builders==1.53.0
binaryornot==0.4.4
blinker==1.9.0
boto3==1.36.12
boto3-stubs==1.35.71
botocore==1.36.12
botocore-stubs==1.36.16
certifi==2025.1.31
cffi==1.17.1
cfn-lint==1.22.7
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cookiecutter==2.6.0
cryptography==44.0.0
dateparser==1.2.1
distlib==0.3.9
docker==7.1.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
mypy-boto3-apigateway==1.35.93
mypy-boto3-cloudformation==1.35.93
mypy-boto3-ecr==1.35.93
mypy-boto3-iam==1.35.93
mypy-boto3-kinesis==1.35.93
mypy-boto3-lambda==1.35.93
mypy-boto3-s3==1.35.93
mypy-boto3-schemas==1.35.93
mypy-boto3-secretsmanager==1.35.93
mypy-boto3-signer==1.35.93
mypy-boto3-sqs==1.35.93
mypy-boto3-stepfunctions==1.35.93
mypy-boto3-sts==1.35.97
mypy-boto3-xray==1.35.93
networkx==3.4.2
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.2
serverless-wsgi==3.0.5
six==1.17.0
sympy==1.13.3
text-unidecode==1.3
tomlkit==0.13.2
types-awscrt==0.23.9
types-python-dateutil==2.9.0.20241206
types-s3transfer==0.11.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
watchdog==4.0.2
Werkzeug==3.1.3
